Move TestCrConfig and TestCrR4Config to main source folder.  Extract TestHapiFhirCrPartitionConfig as separate test config to be used only in the test source folder.  Add Maven compile dependency to hapi-fhir-jpaserver-base in order to ensure that moved Config classes will compile.
