+++ hapi-fhir-storage-cr/pom.xml
@@ -145,6 +145,20 @@
 			<artifactId>jakarta.xml.bind-api</artifactId>
 		</dependency>
 
+		<!-- For test CR config that's exported to downstream modules -->
+		<dependency>
+			<groupId>ca.uhn.hapi.fhir</groupId>
+			<artifactId>hapi-fhir-jpaserver-base</artifactId>
+			<version>${project.version}</version>
+			<exclusions>
+				<!-- prevent conflict with other version of this dependency sourced from elsewhere -->
+				<exclusion>
+					<groupId>org.glassfish</groupId>
+					<artifactId>jakarta.json</artifactId>
+				</exclusion>
+			</exclusions>
+		</dependency>
+
 		<!-- test -->
 		<dependency>
 			<groupId>org.testcontainers</groupId>
