+++ hapi-fhir-storage-cr/src/test/java/ca/uhn/fhir/cr/TestHapiFhirCrPartitionConfig.java
@@ -1,0 +1,10 @@
+package ca.uhn.fhir.cr;
+
+import org.springframework.context.annotation.Bean;
+
+public class TestHapiFhirCrPartitionConfig {
+	@Bean
+	public PartitionHelper partitionHelper() {
+		return new PartitionHelper();
+	}
+}
