+++ hapi-fhir-storage-cr/src/test/java/ca/uhn/fhir/cr/r4/BaseCrR4TestServer.java
@@ -2,10 +2,12 @@ package ca.uhn.fhir.cr.r4;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.cr.IResourceLoader;
+import ca.uhn.fhir.cr.TestHapiFhirCrPartitionConfig;
 import ca.uhn.fhir.cr.config.r4.ApplyOperationConfig;
 import ca.uhn.fhir.cr.config.r4.ExtractOperationConfig;
 import ca.uhn.fhir.cr.config.r4.PackageOperationConfig;
 import ca.uhn.fhir.cr.config.r4.PopulateOperationConfig;
+import ca.uhn.fhir.cr.test.TestCrR4Config;
 import ca.uhn.fhir.jpa.api.config.JpaStorageSettings;
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
 import ca.uhn.fhir.jpa.search.DatabaseBackedPagingProvider;
@@ -38,6 +40,7 @@ import java.util.concurrent.TimeUnit;
 
 
 @ContextConfiguration(classes = {
+	TestHapiFhirCrPartitionConfig.class,
 	TestCrR4Config.class,
 	ApplyOperationConfig.class,
 	ExtractOperationConfig.class,
